requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-environ==0.4.5
gunicorn==20.0.4
psycopg2==2.8.6
whitenoise==5.2.0
django-cors-headers==3.7.0